add test and deploy

+++ division.c
@@ -1,4 +1,4 @@
-float division(int a, int b)
+int division(int a, int b)
 {
     return a / b;
